Added NanoVG as 2D Rendering API First widget, label ! Improved multi-threading
Refactored Demo application code

+++ Pictura/Sources/Maths/Size.h
@@ -3,10 +3,10 @@ namespace Pictura::Maths
 {
 	struct PSize
 	{
-		uint32 Width, Height;
+		float Width, Height;
 
 		PSize() : Width(0), Height(0) {}
-		PSize(uint32 Width, uint32 Height) : Width(Width), Height(Height) {}
+		PSize(float Width, float Height) : Width(Width), Height(Height) {}
 
 		operator PString() const { return "Width[" + Types::ToString(Width) + "] Height[" + Types::ToString(Height) + "]"; }
 		operator PString() { return "Width[" + Types::ToString(Width) + "] Height[" + Types::ToString(Height) + "]"; }
